+++ CHANGES.md
@@ -1,5 +1,5 @@
 # v0.2.7
-+ indels: report length and bq-bin
++ indels: report length and bq-bin and add scripts/analyze-indel-errors.py to plot indel error rate
 + combine_counts: better error messages and handle NA
 
 # v0.2.6